Examples of renderers for o-listpicker and o-combo
