+++ src/app/main/inputs/13.combo/combo-custom-render/combo-custom-render.component.html
@@ -1,0 +1,3 @@
+<ng-template #templateref let-value="value">
+  {{ getComboData(value) }}
+</ng-template>
