+++ src/app/main/inputs/13.combo/input-combo.component.ts
@@ -39,6 +39,40 @@ const COMBO_HTML_DATA_MULTIPLE = `
   </o-form>
 `;
 
+const COMBO_HTML_DATA_CUSTOM_RENDER = `
+  <o-form editable-detail="no" show-header="no" layout-direction="column">
+
+    <o-combo attr="combo-editable-search" [static-data]="getDataArray()" [data]="getValueSimple()"
+      value-column="key" columns="key;value" visible-columns="value" required="yes" read-only="no" null-selection="no" searchable="yes" fxFlex>
+      <combo-custom-render></combo-custom-render>
+    </o-combo>
+
+  </o-form>
+`;
+
+const COMBO_HTML_DATA_BOOLEAN_RENDER = `
+  <o-form editable-detail="no" show-header="no" layout-direction="column">
+
+    <o-combo attr="combo-editable-search" [static-data]="getDataArrayBooleanRenderer()" [data]="getValueBoolean()"
+    value-column="key" columns="key;value" visible-columns="value" required="yes" read-only="no" null-selection="no" searchable="yes" fxFlex>
+      <o-combo-renderer-boolean boolean-type="boolean" render-true-value="Yes" render-false-value="No"></o-combo-renderer-boolean>
+    </o-combo>
+
+  </o-form>
+`;
+
+const COMBO_HTML_DATA_MULTIPLE_CUSTOM_RENDER = `
+  <o-form editable-detail="no" show-header="no" layout-direction="column">
+
+    <o-combo attr="combo-editable-multiple-search" [static-data]="getDataArray()"
+      [data]="getValueMultiple()" value-column="key" columns="key;value" visible-columns="value" required="yes" read-only="no" multiple="yes"
+      null-selection="no" searchable="yes" fxFlex>
+      <combo-custom-render></combo-custom-render>
+    </o-combo>
+  
+  </o-form>
+`;
+
 const COMBO_TS_DATA = `
 import { Component } from '@angular/core';
 
@@ -144,6 +178,106 @@ export class InputComboComponent {
 }
 `;
 
+const COMBO_TS_DATA_CUSTOM_RENDERER = `
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'input-combo',
+  templateUrl: './input-combo.component.html'
+})
+export class InputComboComponent {
+
+  public array: Object[] = [{
+    key: 1,
+    value: '1615472370'
+  }, {
+    key: 2,
+    value: '1515472370'
+  }, {
+    key: 3,
+    value: '1415472370'
+  }, {
+    key: 4,
+    value: '1215472370'
+  }, {
+    key: 5,
+    value: '1115472370'
+  }];
+
+  public getDataArray(): any[] {
+    return this.array;
+  }
+
+  public getValueSimple(): any {
+    return 2;
+  }
+
+}
+`;
+
+const COMBO_TS_DATA_BOOLEAN_RENDERER = `
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'input-combo',
+  templateUrl: './input-combo.component.html'
+})
+export class InputComboComponent {
+
+  public array: Object[] = [{
+    key: 0,
+    value: 0
+  }, {
+    key: 1,
+    value: 1
+  }];
+
+  public getDataArrayBooleanRenderer(): any[] {
+    return this.array;
+  }
+
+  public getValueBoolean(): any {
+    return 1;
+  }
+
+}
+`;
+
+const COMBO_TS_DATA_MULTIPLE_CUSTOM_RENDERER = `
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'input-combo',
+  templateUrl: './input-combo.component.html'
+})
+export class InputComboComponent {
+
+  public multipleValue = [1, 2];
+  public array: Object[] = [{
+    key: 1,
+    value: '1615363293'
+  }, {
+    key: 2,
+    value: '1415363293'
+  }, {
+    key: 3,
+    value: '1215363293'
+  }, {
+    key: 4,
+    value: '1015363293'
+  }];
+
+  public getDataArray(): any[] {
+    return this.array;
+  }
+
+  public getValueMultiple(): any[] {
+    return this.multipleValue;
+  }
+
+}
+`;
+
 @Component({
   selector: 'input-combo',
   templateUrl: './input-combo.component.html'
@@ -183,6 +317,48 @@ export class InputComboComponent {
     value: 'Finland'
   }];
 
+  public array_custom_renderer: Object[] = [{
+    key: 1,
+    value: '1615472370'
+  }, {
+    key: 2,
+    value: '1515472370'
+  }, {
+    key: 3,
+    value: '1415472370'
+  }, {
+    key: 4,
+    value: '1215472370'
+  }, {
+    key: 5,
+    value: '1115472370'
+  }];
+
+  public array_boolean_renderer: Object[] = [{
+    key: 0,
+    value: false
+  }, {
+    key: 1,
+    value: true
+  }];
+
+  public array_date_renderer: Object[] = [{
+    key: 0,
+    value: "10/05/2005"
+  }, {
+    key: 1,
+    value: "04/03/1980"
+  }, {
+    key: 2,
+    value: "24/08/2011"
+  }, {
+    key: 3,
+    value: "01/12/1980"
+  }, {
+    key: 4,
+    value: "30/01/2016"
+  }];
+
   public files = {
     html: {
       data: COMBO_HTML_DATA
@@ -207,12 +383,64 @@ export class InputComboComponent {
     }
   };
 
+  public files_custom_render = {
+    html: {
+      data: COMBO_HTML_DATA_CUSTOM_RENDER
+    },
+    scss: {
+      data: undefined
+    },
+    typescript: {
+      data: COMBO_TS_DATA_CUSTOM_RENDERER
+    }
+  };
+
+  public files_boolean_render = {
+    html: {
+      data: COMBO_HTML_DATA_BOOLEAN_RENDER
+    },
+    scss: {
+      data: undefined
+    },
+    typescript: {
+      data: COMBO_TS_DATA_BOOLEAN_RENDERER
+    }
+  };
+
+  public files_multiple_custom_render = {
+    html: {
+      data: COMBO_HTML_DATA_MULTIPLE_CUSTOM_RENDER
+    },
+    scss: {
+      data: undefined
+    },
+    typescript: {
+      data: COMBO_TS_DATA_MULTIPLE_CUSTOM_RENDERER
+    }
+  };
+
   public getDataArray(): any[] {
     return this.array;
   }
 
+  public getDataArrayCustomRenderer(): any[] {
+    return this.array_custom_renderer;
+  }
+
+  public getDataArrayBooleanRenderer(): any[] {
+    return this.array_boolean_renderer;
+  }
+
+  public getDataArrayDateRenderer(): any[] {
+    return this.array_date_renderer;
+  }
+
   public getValueSimple(): any {
     return 2;
+  }
+
+  public getValueBoolean(): any {
+    return 1;
   }
 
   public getValueMultiple(): any[] {
