+++ src/app/main/inputs/14.listpicker/input-listpicker.component.html
@@ -1,0 +1,96 @@
+<example-comp comp-name="INPUTS.LISTPICKER" comp-desc="INPUTS.LISTPICKER.DESC" [files]="files">
+  <o-form editable-detail="false" show-header="no" layout-direction="column">
+    <div fxLayout="column" layout-padding>
+      <label class="input-comp-title">{{ 'INPUTS.READ_ONLY' | oTranslate }}</label>
+      <o-list-picker attr="listpicker" label="{{ 'INPUT.BUTTON.LISTPICKER' | oTranslate }}" [static-data]="getDataArray()" [data]="getValue()"
+        filter="yes" value-column="key" columns="key;value" visible-columns="value"></o-list-picker>
+    </div>
+    <div fxLayout="column" layout-padding>
+      <label class="input-comp-title">{{ 'INPUTS.EDITABLE' | oTranslate }}</label>
+      <o-list-picker #listpicker attr="listpicker2" label="{{ 'INPUT.BUTTON.LISTPICKER' | oTranslate }} - Custom Render" [static-data]="getDataArray()"
+        [data]="getValue()" filter="yes" value-column="key" columns="key;value" visible-columns="value" required="true" read-only="false">
+      </o-list-picker>
+    </div>
+    <div fxLayout="column" layout-padding>
+      <label class="input-comp-title">{{ 'INPUTS.DISABLED' | oTranslate }}</label>
+      <o-list-picker attr="listpicker3" label="{{ 'INPUT.BUTTON.LISTPICKER' | oTranslate }}" [static-data]="getDataArray()" [data]="getValue()"
+        enabled="no" filter="yes" value-column="key" columns="key;value" visible-columns="value"></o-list-picker>
+    </div>
+  </o-form>
+</example-comp>
+
+<example-comp comp-name="INPUTS.LISTPICKER_CUSTOM_RENDERER" comp-desc="INPUTS.LISTPICKER_CUSTOM_RENDERER.DESC" [files]="files_custom_render">
+  <o-form editable-detail="false" show-header="no" layout-direction="column">
+    <div fxLayout="column" layout-padding>
+      <label class="input-comp-title">{{ 'INPUT.BUTTON.LISTPICKER' | oTranslate }}</label>
+      <o-list-picker #listpicker attr="listpicker2" label="{{ 'INPUT.BUTTON.LISTPICKER_CUSTOM_RENDERER' | oTranslate }}" [static-data]="getDataArrayCustomRenderer()"
+        [data]="getValue()" filter="yes" value-column="key" columns="key;value" visible-columns="value" required="true" read-only="false">
+        <listpicker-custom-render></listpicker-custom-render>
+      </o-list-picker>
+    </div>
+  </o-form>
+</example-comp>
+
+
+<example-comp comp-name="INPUTS.LISTPICKER_CURRENCY_RENDERER" comp-desc="INPUTS.LISTPICKER_CURRENCY_RENDERER.DESC" [files]="files_currency_render">
+  <o-form editable-detail="false" show-header="no" layout-direction="column">
+    <div fxLayout="column" layout-padding>
+      <label class="input-comp-title">{{ 'INPUT.BUTTON.LISTPICKER' | oTranslate }}</label>
+      <o-list-picker #listpicker attr="listpicker2" label="{{ 'INPUT.BUTTON.LISTPICKER_WITH_RENDERER' | oTranslate }}" [static-data]="getDataArrayCustomRenderer()"
+        [data]="getValue()" filter="yes" value-column="key" columns="key;value" visible-columns="value" required="true" read-only="false">
+        <o-listpicker-renderer-currency></o-listpicker-renderer-currency>
+      </o-list-picker>
+    </div>
+  </o-form>
+</example-comp>
+
+
+<example-comp comp-name="INPUTS.LISTPICKER_REAL_RENDERER" comp-desc="INPUTS.LISTPICKER_REAL_RENDERER.DESC" [files]="files_real_render">
+  <o-form editable-detail="false" show-header="no" layout-direction="column">
+    <div fxLayout="column" layout-padding>
+      <label class="input-comp-title">{{ 'INPUT.BUTTON.LISTPICKER' | oTranslate }}</label>
+      <o-list-picker #listpicker attr="listpicker2" label="{{ 'INPUT.BUTTON.LISTPICKER_WITH_RENDERER' | oTranslate }}" [static-data]="getDataArrayCustomRenderer()"
+        [data]="getValue()" filter="yes" value-column="key" columns="key;value" visible-columns="value" required="true" read-only="false">
+        <o-listpicker-renderer-real></o-listpicker-renderer-real>
+      </o-list-picker>
+    </div>
+  </o-form>
+</example-comp>
+
+
+<example-comp comp-name="INPUTS.LISTPICKER_INTEGER_RENDERER" comp-desc="INPUTS.LISTPICKER_INTEGER_RENDERER.DESC" [files]="files_integer_render">
+  <o-form editable-detail="false" show-header="no" layout-direction="column">
+    <div fxLayout="column" layout-padding>
+      <label class="input-comp-title">{{ 'INPUT.BUTTON.LISTPICKER' | oTranslate }}</label>
+      <o-list-picker #listpicker attr="listpicker2" label="{{ 'INPUT.BUTTON.LISTPICKER_WITH_RENDERER' | oTranslate }}" [static-data]="getDataArrayCustomRenderer()"
+        [data]="getValue()" filter="yes" value-column="key" columns="key;value" visible-columns="value" required="true" read-only="false">
+        <o-listpicker-renderer-integer></o-listpicker-renderer-integer>
+      </o-list-picker>
+    </div>
+  </o-form>
+</example-comp>
+
+<example-comp comp-name="INPUTS.LISTPICKER_PERCENTAGE_RENDERER" comp-desc="INPUTS.LISTPICKER_PERCENTAGE_RENDERER.DESC" [files]="files_percentage_render">
+  <o-form editable-detail="false" show-header="no" layout-direction="column">
+    <div fxLayout="column" layout-padding>
+      <label class="input-comp-title">{{ 'INPUT.BUTTON.LISTPICKER' | oTranslate }}</label>
+      <o-list-picker #listpicker attr="listpicker2" label="{{ 'INPUT.BUTTON.LISTPICKER_WITH_RENDERER' | oTranslate }}" [static-data]="getDataArrayPercentageRenderer()"
+        [data]="getValuePercentage()" filter="yes" value-column="key" columns="key;value" visible-columns="value" required="true" read-only="false">
+        <o-listpicker-renderer-percentage value-base="100"></o-listpicker-renderer-percentage>
+      </o-list-picker>
+    </div>
+  </o-form>
+</example-comp>
+
+
+<example-comp comp-name="INPUTS.LISTPICKER_DATE_RENDERER" comp-desc="INPUTS.LISTPICKER_DATE_RENDERER.DESC" [files]="files_date_render">
+  <o-form editable-detail="false" show-header="no" layout-direction="column">
+    <div fxLayout="column" layout-padding>
+      <label class="input-comp-title">{{ 'INPUT.BUTTON.LISTPICKER' | oTranslate }}</label>
+      <o-list-picker #listpicker attr="listpicker" label="{{ 'INPUT.BUTTON.LISTPICKER_WITH_RENDERER' | oTranslate }}" [static-data]="getDataArrayDateRenderer()"
+        [data]="getValue()" filter="yes" value-column="key" columns="key;value" visible-columns="value" required="true" read-only="false">
+        <o-listpicker-renderer-date format="YYYY-MM-DD"></o-listpicker-renderer-date>
+      </o-list-picker>
+    </div>
+  </o-form>
+</example-comp>
