+++ src/app/main/inputs/15.html/input-html.component.html
@@ -1,0 +1,12 @@
+<example-comp comp-name="INPUTS.HTML" comp-desc="INPUTS.HTML.DESC" [files]="files">
+  <o-form editable-detail="no" show-header="no" layout-direction="column">
+    <div fxLayout="column" layout-padding>
+      <label class="input-comp-title">{{ 'INPUTS.READ_ONLY' | oTranslate }}</label>
+      <o-html-input #input1 attr="html" [data]="getValue()"></o-html-input>
+    </div>
+    <div fxLayout="column" layout-padding>
+      <label class="input-comp-title">{{ 'INPUTS.EDITABLE' | oTranslate }}</label>
+      <o-html-input attr="html2" [data]="getValue()" read-only="no" required="yes"></o-html-input>
+    </div>
+  </o-form>
+</example-comp>
