+++ src/assets/i18n/en.json
@@ -144,7 +144,8 @@
   "INPUT.BUTTON.FILE": "File",
   "INPUT.BUTTON.CHECKBOX": "Checkbox",
   "INPUT.BUTTON.LISTPICKER": "List picker",
-  "INPUT.BUTTON.LISTPICKER_CUSTOM_RENDERER": "List picker Custom Renderer",
+  "INPUT.BUTTON.LISTPICKER_CUSTOM_RENDERER": "List picker with custom renderer",
+  "INPUT.BUTTON.LISTPICKER_WITH_RENDERER": "List picker with renderer",
   "INPUT.BUTTON.VALIDATORS": "Validators",
   "INPUT.BUTTON.COMBO": "Combo",
   "INPUT.BUTTON.COMBO_CUSTOM_RENDERER": "Combo Custom Renderer",
@@ -202,7 +203,7 @@
   "INPUTS.CHECKBOXBOOLEANTYPE_BOOLEAN": "Boolean",
   "INPUTS.CHECKBOXBOOLEANTYPE.DESC": "There are 3 boolean types (number, string and boolean) ",
   "INPUTS.COMBO": "Combo inputs",
-  "INPUTS.COMBO_CUSTOM_RENDERER": "Combo inputs custom renderer",
+  "INPUTS.COMBO_CUSTOM_RENDERER": "Combo input with custom renderer",
   "INPUTS.COMBO_MULTIPLE": "Combo inputs (multiple selection)",
   "INPUTS.COMBO_MULTIPLE_CUSTOM_RENDERER": "Combo inputs (multiple selection) with custom renderer",
   "INPUTS.COMBO.DESC": "A component for getting or displaying an option from multiple.",
@@ -210,9 +211,19 @@
   "INPUTS.COMBO.DESC_MULTIPLE": "A combo component that allows multiple selection",
   "INPUTS.COMBO_CUSTOM_RENDERER.DESC_MULTIPLE": "A combo component that allows multiple selection with custom renderer",
   "INPUTS.LISTPICKER": "List picker inputs",
-  "INPUTS.LISTPICKER_CUSTOM_RENDERER": "List picker custom renderer",
+  "INPUTS.LISTPICKER_CUSTOM_RENDERER": "List picker with custom renderer",
+  "INPUTS.LISTPICKER_REAL_RENDERER": "List picker with real renderer",
+  "INPUTS.LISTPICKER_INTEGER_RENDERER": "List picker with integer renderer",
+  "INPUTS.LISTPICKER_CURRENCY_RENDERER": "List picker with currency renderer",
+  "INPUTS.LISTPICKER_PERCENTAGE_RENDERER": "List picker with percentage renderer",
+  "INPUTS.LISTPICKER_DATE_RENDERER": "List picker with date renderer",
   "INPUTS.LISTPICKER.DESC": "A component for getting or displaying an option from multiple.",
   "INPUTS.LISTPICKER_CUSTOM_RENDERER.DESC": "A component for getting or displaying an option from multiple with your own renderer.",
+  "INPUTS.LISTPICKER_REAL_RENDERER.DESC": "A component for getting or displaying an option from multiple with real renderer.",
+  "INPUTS.LISTPICKER_INTEGER_RENDERER.DESC": "A component for getting or displaying an option from multiple with integer renderer.",
+  "INPUTS.LISTPICKER_CURRENCY_RENDERER.DESC": "A component for getting or displaying an option from multiple with currency renderer.",
+  "INPUTS.LISTPICKER_PERCENTAGE_RENDERER.DESC": "A component for getting or displaying an option from multiple with percentage renderer.",
+  "INPUTS.LISTPICKER_DATE_RENDERER.DESC": "A component for getting or displaying an option from multiple with date renderer.",
   "INPUTS.HOUR": "Hour",
   "INPUTS.HOUR.DESC": "A component for getting or displaying hour submitted by the user.",
   "INPUTS.HOUR.FORMATS": "Formats",
