+++ src/assets/i18n/es.json
@@ -140,7 +140,8 @@
   "INPUT.BUTTON.FILE": "Fichero",
   "INPUT.BUTTON.CHECKBOX": "Checkbox",
   "INPUT.BUTTON.LISTPICKER": "List picker",
-  "INPUT.BUTTON.LISTPICKER_CUSTOM_RENDERER": "List picker - Renderer personalizado",
+  "INPUT.BUTTON.LISTPICKER_CUSTOM_RENDERER": "List picker con renderer personalizado",
+  "INPUT.BUTTON.LISTPICKER_WITH_RENDERER": "List picker con renderer",
   "INPUT.BUTTON.VALIDATORS": "Validadores",
   "INPUT.BUTTON.COMBO": "Combo",
   "INPUT.BUTTON.COMBO_CUSTOM_RENDERER": "Combo - Renderer personalizado",
@@ -204,9 +205,19 @@
   "INPUTS.COMBO.DESC_MULTIPLE": "Componente que permite selección múltiple",
   "INPUTS.COMBO_CUSTOM_RENDERER.DESC_MULTIPLE": "Componente que permite selección múltiple con renderer personalizado",
   "INPUTS.LISTPICKER": "Inputs list picker",
-  "INPUTS.LISTPICKER_CUSTOM_RENDERER": "List picker renderer personalizado",
+  "INPUTS.LISTPICKER_CUSTOM_RENDERER": "List picker con renderer personalizado",
+  "INPUTS.LISTPICKER_REAL_RENDERER": "List picker con renderer real",
+  "INPUTS.LISTPICKER_INTEGER_RENDERER": "List picker con renderer entero",
+  "INPUTS.LISTPICKER_CURRENCY_RENDERER": "List picker con renderer de tipo moneda",
+  "INPUTS.LISTPICKER_PERCENTAGE_RENDERER": "List picker con renderer de porcentaje",
+  "INPUTS.LISTPICKER_DATE_RENDERER": "List picker con renderer de fecha",
   "INPUTS.LISTPICKER.DESC": "Componente para insertar o mostrar una opción entre múltiples.",
   "INPUTS.LISTPICKER_CUSTOM_RENDERER.DESC": "Componente para insertar o mostrar una opción entre múltiples con renderer personalizado.",
+  "INPUTS.LISTPICKER_REAL_RENDERER.DESC": "Componente para insertar o mostrar una opción entre múltiples con renderer de tipo real.",
+  "INPUTS.LISTPICKER_INTEGER_RENDERER.DESC": "Componente para insertar o mostrar una opción entre múltiples con renderer de tipo entero.",
+  "INPUTS.LISTPICKER_CURRENCY_RENDERER.DESC": "Componente para insertar o mostrar una opción entre múltiples con renderer de tipo moneda.",
+  "INPUTS.LISTPICKER_PERCENTAGE_RENDERER.DESC": "Componente para insertar o mostrar una opción entre múltiples con renderer de tipo porcentaje.",
+  "INPUTS.LISTPICKER_DATE_RENDERER.DESC": "Componente para insertar o mostrar una opción entre múltiples con renderer de tipo fecha.",
   "INPUTS.HOUR": "Hora",
   "INPUTS.HOUR.DESC": "Componente para insertar o mostrar la hora.",
   "INPUTS.HOUR.FORMATS": "Formatos",
